refactor: Integrado módulo count en el módulo voting #50
