+++ decide/decide/settings.py
@@ -63,7 +63,6 @@ MODULES = [
     'base',
     'booth',
     'census',
-    'count',
     'mixnet',
     'postproc',
     'store',
